requirements.txt
# Dependencias para Ofertix Backend - VERSIÓN REAL
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Base de datos
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7  # Para PostgreSQL

# Autenticación y seguridad
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3

# Pagos
stripe==6.6.0

# APIs y scraping
selenium==4.11.2
webdriver-manager==4.0.0
beautifulsoup4==4.12.2

# Utilidades
celery==5.3.2  # Para tareas asíncronas
redis==5.0.0   # Para cache y cola de tareas
gunicorn==21.2.0  # Para producción

# Monitoreo
flask-cors==4.0.0
# selenium==4.11.2
# webdriver-manager==4.0.0